requirements.txt
aiohttp==3.8.4
browser_cookie3==0.19.1
colorama==0.4.5
deep_translator==1.11.4
docx==0.2.4
httpx==0.25.2
langdetect==1.0.9
Markdown==3.5.1
pandas==2.1.4
protobuf==4.25.1
python_docx==1.1.0
PyYAML==3.13
Requests==2.31.0
setuptools==65.5.1
streamlit==1.29.0
google-generativeai
pypdf
